works, but not really

+++ src/se/kth/networking/java/first/ApplicationDomain.java
@@ -1,10 +1,12 @@
 package se.kth.networking.java.first;
 
+import java.math.BigInteger;
+
 /**
  * Created by victoraxelsson on 2016-11-18.
  */
 public interface ApplicationDomain {
-    void storeKey(long key, String value);
-    String getKey(long key);
-    void foundKey(long key, String value);
+    void storeKey(BigInteger key, String value);
+    String getKey(BigInteger key);
+    void foundKey(BigInteger key, String value);
 }
